### 2019-05-02 #1424 @kyleniemeyer
@mommermi I just submitted a PR (NASA-Planetary-Science/sbpy#143) that should fix the paper compilation issue.
