+++ publications/joss/paper.md
@@ -112,7 +112,7 @@ C-extensions are utilized.
 Designed as an ``Astropy`` affiliated package, ``sbpy`` utilizes a
 wide range of functionality of ``Astropy``, including its table,
 constants, units, and modeling submodules, and other affiliated
-packages like astroquery [@astroquery]. ``sbpy``'s API is deliberately emulating
+packages like ``astroquery`` [@astroquery]. ``sbpy``'s API is deliberately emulating
 that of ``Astropy`` wherever possible in order to enable user-friendly
 and consistent access to its functionality.
 
