### 2019-05-02 b338732
We rely on sphninx-astropy rather than on the helpers for the plugins
