+++ requirements.txt
@@ -1,5 +1,4 @@
 numpy>=1.4.0
-astropy-helpers
 matplotlib
 ads
 synphot
