### 2019-05-02 55cdfc0
Minor template updates
