### 2019-05-03 #1426 @whedon
[REVIEW]: sbpy: A Python module for small-body planetary astronomy
**Submitting author:** @mommermi (<a href="http://orcid.org/0000-0002-8132-778X">Michael Mommert</a>)
**Repository:** <a href="https://github.com/NASA-Planetary-Science/sbpy" target ="_blank">https://github.com/NASA-Planetary-Science/sbpy</a>
**Version:** v0.1.1
