### 2019-05-03 #1426 @Juanlu001
Before I finish a proper review:

> Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

The license is stated in the README and also in https://github.com/NASA-Planetary-Science/sbpy/blob/master/licenses/LICENSE.rst. Not what I expected, but I guess it should be enough.
