### 2019-05-03 #1426 @mommermi
Hi @Juanlu001, thanks for reviewing this submission! Would you prefer the `License.rst` to be on the root level (probably makes more sense on the root level anyway...)?
