### 2019-05-03 #1426 @Juanlu001
I think so @mommermi! Having a `LICENSE` at the top (not sure about the extensions) makes it easier to spot and also GitHub can read the metadata from it.
