+++ licenses/README.rst
@@ -3,3 +3,6 @@ Licenses
 
 This directory holds license and credit information for the affiliated package,
 works the affiliated package is derived from, and/or datasets.
+
+The license file for the sbpy package itself is placed in the root directory
+of this repository.
