### 2019-05-03 #1426 @mommermi
Ok, done!
I created a PR (NASA-Planetary-Science/sbpy#145) that will hold all the code changes from this review.
