### 2019-05-03 f41f87d
added Brigitta to author list
