### 2019-05-03 716f610
all models implemented, H-d-pv conversion implemented
