+++ sbpy/thermal/core.py
@@ -21,7 +21,7 @@ from ..data import Ephem, Phys
 from ._thermal import integrate_planck
 
 
-__all__ = ['ThermalClass', 'STM', 'FRM', 'NEATM']
+__all__ = ['ThermalClass', 'STM', 'FRM', 'NEATM', 'NEATM_fixedeta']
 
 
 class ThermalWarning(AstropyUserWarning):
@@ -40,7 +40,8 @@ class ThermalClass(Fittable1DModel):
     :math:`\eta` can be derived from the latter.
 
     All thermal models are derived from this class: `~sbpy.thermal.STM`,
-    `~sbpy.thermal.FRM`, and `~sbpy.thermal.NEATM`. The actual model
+    `~sbpy.thermal.FRM`, `~sbpy.thermal.NEATM`, and
+    `~sbpy.thermal.NEATM_fixedeta`. The actual model
     model implementations can be found as C extensions in
     `~sbpy.thermal._thermal`; the implementations are accessible through
     `~sbpy.thermal._thermal.integrate_planck`.
@@ -48,9 +49,10 @@ class ThermalClass(Fittable1DModel):
 
     # define astropy.modeling.Fittable1DModel model parameters:
     # target diameter
-    diam = Parameter(default=1, min=0, unit=u.km)
-    # subsolar temperature normalized to heliocentric distance of 1 au
-    subsolartemp_au = Parameter(default=150, min=0, unit=u.K)
+    diam = Parameter(default=1, min=1e-6, unit=u.km)
+    # target Bondalbedo
+    bondalbedo = Parameter(default=0.05, min=1e-5, max=1,
+                           unit=u.dimensionless_unscaled)
 
     # enable dimensionless input parameter (wavelength)
     _input_units_allow_dimensionless = True
@@ -65,8 +67,9 @@ class ThermalClass(Fittable1DModel):
 
     def _parameter_units_for_data_units(self, inputs_unit, outputs_unit):
         """Defines units of model parameters."""
-        return OrderedDict([('subsolartemp_au', u.K),
-                            ('diam', u.km)])
+        return OrderedDict([('bondalbedo', u.dimensionless_unscaled),
+                            ('diam', u.km),
+                            ])
 
     def __init__(self, *pargs):
         """A `ThermalClass` object can be instantiated with or without
@@ -98,8 +101,9 @@ class ThermalClass(Fittable1DModel):
             ``phaseangle`` (solar phase angle), ``wavelength``
             (wavelength of observation), ``flux`` (measured flux
             density at a given wavelength), ``fluxerr`` (error of the
-            flux density measurement), and ``subsolartemp`` (subsolar
-            temperature). Refer to the documentation of methods for
+            flux density measurement), ``subsolartemp`` (subsolar
+            temperature), and ``eta`` (infrared beaming factor). Refer to
+            the documentation of methods for
             flux estimation and thermal model fitting for lists of
             required fields.
 
@@ -165,11 +169,11 @@ class ThermalClass(Fittable1DModel):
         -----
         This method requires the following fields in ``self.phys``:
            * ``bondalbedo``: target's Bond albedo
-           * ``eta``: infrared beaming parameter (see Harris 1998)
            * ``emissivity``: target emissivity
 
         This method requires the following fields in ``self.ephem``:
            * ``heliodist``: heliocentric distance of the target
+           * ``eta``: infrared beaming parameter (see Harris 1998)
 
         Returns
         -------
@@ -180,7 +184,7 @@ class ThermalClass(Fittable1DModel):
         subsolartemp = (const.L_sun/(4*np.pi*const.au.to(u.m)**2) *
                         (1.-self.phys['bondalbedo']) /
[…]
