### 2019-05-03 a8b5c9f
implemented albedo conversions
