+++ sbpy/thermal/core.py
@@ -435,104 +435,181 @@ class ThermalClass(Fittable1DModel):
         else:
             return absmag
 
-    # def calculate_pv_from_bondalbedo(self, append_results=False):
-    #     """Calculate the geometric albedo and the Bond albedo based on
-    #     the target's physical properties.
-
-    #     Parameters
-    #     ----------
-    #     append_results : bool, optional
-    #         If `append_results=True`, results from this function are
-    #         appended to `self.phys` and a copy of `self.phys` is returned;
-    #         if `append_results=False`,
-    #         the geometric albedo and the Bond albedo are returned as as
-    #         a separate `~sbpy.data.Phys` object. Default: `False`
-
-    #     Notes
-    #     -----
-    #     This method requires the following fields in ``self.phys``:
-    #        * ``diam``: target's diameter
-    #        * ``absmag``: target's absolute magnitude (V-band)
-    #        * ``slopepar``: target's photometric slope parameter TBD
-
-    #     Returns
-    #     -------
-    #     `~sbpy.data.Phys` object.
-    #     """
-
-    #     pv = (self.phys['bondalbedo'][0] /
-    #           (0.29+0.684*self.phys['G'][0]))
-
-    #     if append_results:
-    #         self.phys['pv'] = pv
-    #         return Phys.from_table(self.phys.table)
-    #     else:
-    #         return Phys.from_dict({'pv': pv}),
-
-    #     if all([f in self.phys.column_names for f in ['G', 'pv']):
-
-    #     if 'H' in self.phys.column_names:
-
-    #     # ref goes here
-    #     pv = ((1329/self.phys['diam'].to('km').value *
-    #            10**(-self.phys['H'].to('mag').value/5))**2)
-
-    #     bondalbedo_from_pv = pv*(0.29+0.684*self.phys['G'][0])
-
-    #     print('bondalbedos', self.phys['bondalbedo'][0], bondalbedo_from_pv)
-
-    #     print('pvs', pv, pv_from_bondalbedo)
-    #     if append_results:
-    #         self.phys['pv'] = pv
-    #         return Phys.from_table(self.phys.table)
-    #     else:
-    #         return Phys.from_dict({'pv': pv}),
-    #         # 'bondalbedo': bondalbedo})
-
-    def eta_from_subsolartemp(self, append_results=False):
-        """Calculate the beaming parameter :math:`\eta` from the target's
-        subsolar temperature, its physical properties, and observational
-        circumstances.
+    def calculate_pv_from_bondalbedo(self, append_results=False, **kwargs):
+        """Calculate the target's geometric albedo (V-band) from its Bond
+        albedo.
 
         Parameters
         ----------
         append_results : bool, optional
             If `append_results=True`, results from this function are
             appended to `self.phys` and a copy of `self.phys` is returned;
-            if `append_results=False`, the beaming parameter :math:`\eta`
-            is returned as as a `~astropy.units.Quantity` object. Default:
-            `False`
+            if `append_results=False`,
+            the geometric albedo and the Bond albedo are returned as as
+            a separate `~sbpy.data.Phys` object. Default: `False`
 
         Notes
         -----
-        This method requires the following fields in ``self.phys``:
+        This method requires the following field in ``self.phys``:
            * ``bondalbedo``: target's Bond albedo
-           * ``emissivity``: target's emissivity
 
-        This method requires the following fields in ``self.ephem``:
-           * ``subsolartemp``: target's subsolar temperature per epoch
-           * ``heliodist``: target's heliocentric distance per epoch
+        The photometric slop parameter ``G`` can be provided; if it is not
+        provided it defaults to ``G=0.15``.
+
+        Alternatively, these properties can also be provided as
+        keyword arguments (see example below). In that case, both
[…]
