+++ docs/fieldnames.rst
@@ -20,7 +20,8 @@ As an example, heliocentric distance can be addressed a ``'r'`` or
 
 The list of alternative field names is always up to date, but not
 complete. The source list is located as
-``sbpy.data.conf.fieldnames``. If you think an important alternative
+``sbpy.data.conf.fieldnames`` in file ``sbpy/data/__init__.py``. If you
+think an important alternative
 is missing, please suggest it by opening an issue. However, keep in mind
 that each alternative field name has to be *unique* and *unambiguous*.
 
@@ -31,9 +32,9 @@ List of Alternative Field Names
 =================================== ===================================================================================================================
                         Description                                                                                                   Alternative Names
 =================================== ===================================================================================================================
-              **Target Identifier**                                                                                              ``targetname``, ``id``
+              **Target Identifier**                                                                                  ``targetname``, ``id``, ``Object``
                     **Inclination**                                                                                            ``i``, ``inc``, ``incl``
-                          **Epoch**                                                                      ``epoch``, ``datetime_jd``, ``Date``, ``date``
+                          **Epoch**                                                              ``epoch``, ``datetime_jd``, ``JD``, ``Date``, ``date``
 **Longitude of the Ascending Node**                                                                                             ``Omega``, ``longnode``
       **Argument of the Periapsis**                                                                                                   ``w``, ``argper``
           **Heliocentric Distance**                                                                               ``r``, ``rh``, ``r_hel``, ``heldist``
@@ -41,21 +42,22 @@ List of Alternative Field Names
                 **Right Ascension**                                                                                                      ``ra``, ``RA``
                     **Declination**                                                                                           ``dec``, ``DEC``, ``Dec``
                         **RA Rate**                                              ``ra_rate``, ``RA_rate``, ``ra_rates``, ``RA_rates``, ``dRA``, ``dra``
-               **RA*cos(Dec) Rate**                        ``RA*cos(Dec)_rate``, ``dra*cos(dec)``, ``dRA*cos(Dec)``, ``dra cos(dec)``, ``dRA cos(Dec)``
+               **RA*cos(Dec) Rate**                        ``RA*cos(Dec)_rate``, ``dra cos(dec)``, ``dRA cos(Dec)``, ``dra*cos(dec)``, ``dRA*cos(Dec)``
                        **Dec Rate** ``dec_rate``, ``DEC_rate``, ``Dec_rate``, ``dec_rates``, ``DEC_rates``, ``Dec_rates``, ``dDec``, ``dDEC``, ``ddec``
                   **Proper Motion**                                                                                           ``mu``, ``Proper motion``
-        **Proper Motion Direction**                                                                                        ``direction``, ``Direction``
+        **Proper Motion Direction**                                                                                         ``Direction``, ``directin``
               **Solar Phase Angle**                                                                                ``alpha``, ``phaseangle``, ``Phase``
                **Solar Elongation**                                      ``elong``, ``solarelong``, ``solarelongation``, ``elongation``, ``Elongation``
                **V-band Magnitude**                                                                                                     ``V``, ``Vmag``
 **Heliocentric Ecliptic Longitude**                                                      ``hlon``, ``EclLon``, ``ecllon``, ``HelEclLon``, ``helecllon``
  **Heliocentric Ecliptic Latitude**                                                      ``hlat``, ``EclLat``, ``ecllat``, ``HelEclLat``, ``helecllat``
                       **Elevation**                                                  ``el``, ``EL``, ``elevation``, ``alt``, ``altitude``, ``Altitude``
-		        **Azimuth**                                                                                         ``az``, ``AZ``, ``azimuth``
[…]
