### 2019-05-07 679aa18
misc improvements to docstrings and code
In response to code review by MSP Kelley
