### 2019-05-07 5ed6926
misc improvements to docstrings and code
In response to code review by MSP Kelley
