### 2019-05-08 0f12e5d
update docstring
