### 2019-05-08 8777a04
bugfix units/tests/test_core.py
