### 2019-05-08 9784629
Revert "try underscore instead of hyphen in file names"
This reverts commit 84a44d14534f13ab7512117ca5cdd8d44a735604.
