### 2019-05-31 9867a86
Correct typo in docs
