+++ sbpy/spectroscopy/reflectance/__init__.py
@@ -1,0 +1,9 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+=============================================
+sbpy Asteroid Reflectance Spectroscopy Module
+=============================================
+created on June 4, 2019
+"""
+
+from .core import *
