### 2019-06-06 ec9b068
reflectance spectrum model fitting implemented
