+++ sbpy/spectroscopy/reflectance/core.py
@@ -15,7 +15,7 @@ from collections import OrderedDict
 from scipy.interpolate import InterpolatedUnivariateSpline
 
 import astropy.units as u
-from astropy.modeling import Fittable1DModel, Parameter
+from astropy.modeling import Fittable1DModel, Parameter, fitting
 from astropy.utils.data import get_pkg_data_filename
 from astropy.table import Table
 
@@ -33,7 +33,7 @@ class Taxon(Fittable1DModel):
     @property
     def input_units(self):
         """Define units of input data."""
-        return {'x': u.micron}
+        return {'x': u.um}
 
     def _parameter_units_for_data_units(self, inputs_unit, outputs_unit):
         """Define units of model parameters."""
@@ -42,67 +42,58 @@ class Taxon(Fittable1DModel):
     def __init__(self, taxon, schema='BusDeMeo', slope=0):
         self.schema = schema
         self.taxon = taxon
-        self.slope = slope
 
         self.raw_spec = None  # raw discrete spectrum
         self.spec = None  # interpolated spectrum
         self.sigma = None  # interpolated (nearest-neighbor) uncertainties
 
+        self.spline_order = None  # polynomial order used for splines
         self.wavelength_range = self.from_file(schema, taxon)
 
-    def from_file(self, schema, taxon, spline_order=3):
+    def from_file(self, schema, taxon, reddening_slope=0*u.percent/u.um,
+                  spline_order=3):
 
         from astropy.utils.data import _is_url
 
         try:
             parameters = getattr(schemas, schema).copy()
-
-            # locate data file
-            if not _is_url(parameters['filename']):
-                # find in the module's location
-                parameters['filename'] = get_pkg_data_filename(
-                    os.path.join('data', parameters['filename']))
-
-            # read data file
-            spec = Table.read(parameters['filename'],
-                              format='ascii.commented_header',
-                              header_start=2,
-                              delimiter=' ',
-                              fill_values=[('-0.999', '0')])
-
-            # extract spectrum for this taxon
-            spec = spec['Wavelength', taxon+'_Mean', taxon+'_Sigma']
-            spec.rename_column(taxon+'_Mean', 'Spec')
-            spec.rename_column(taxon+'_Sigma', 'Sigma')
-            spec['Wavelength'].unit = parameters['wave_unit']
-            spec['Spec'].unit = parameters['flux_unit']
-            spec['Sigma'].unit = parameters['flux_unit']
-            self.raw_spec = Phys.from_table(spec)
-
-            # # apply reddening
-            # if reddening_slope > 0:
-            #     self.raw_spec = self.redden(
-            #         self.raw_spec, reddening_slope)
-
-            # derive weights limiting minimum uncertainties
-            weights = np.array([1/max(val, 0.001)**2 for val in
-                                self.raw_spec['Sigma'].value])
-
-            # interpolate spectrum using splines
-            self.spec = InterpolatedUnivariateSpline(
-                self.raw_spec['Wavelength'].to('um').value,
-                self.raw_spec['Spec'].value,
-                w=weights,
-                k=spline_order)
-
         except AttributeError:
             msg = 'Unknown taxonomy schema "{}".  Valid schemas:\n{}'.format(
                 schema, schemas.available)
             raise ValueError(msg)
+
+        # locate data file
+        try:
+            if not _is_url(parameters['filename']):
+                # find in the module's location
+                parameters['filename'] = get_pkg_data_filename(
+                    os.path.join('data', parameters['filename']))
[…]
