### 2019-06-07 7025c4d
implemented offset parameter for fitting
