### 2019-06-09 NASA-Planetary-Science/sbpy#149 @Juanlu001
Orbit example from docs can't be propagated
I tried the first `Orbit` example from the docs and the propagation raises an error:

```
In [1]: %cpaste
Pasting code; enter '--' alone on the line to stop or use Ctrl-D.
:>>> from sbpy.data import Orbit
>>> import astropy.units as u
>>> elements = {'a':1.234*u.au, 'e':0.1234, 'i':12.34*u.deg,
...             'argper': 123.4*u.deg, 'node': 45.2*u.deg,
...             'epoch': 2451200.5*u.d, 'true_anom':23.1*u.deg}
>>> orb = Orbit.from_dict(elements)
>>> print(orb)  # doctest: +SKIP::::::
:<EOF>
<QTable length=1>
   a       e       i     argper   node    epoch   true_anom
   AU             deg     deg     deg       d        deg
float64 float64 float64 float64 float64  float64   float64
------- ------- ------- ------- ------- --------- ---------
  1.234  0.1234   12.34   123.4    45.2 2451200.5      23.1

In [2]: from astropy.time import Time

In [3]: orb.oo_propagate(Time.now())
KeyError                                  Traceback (most recent call last)
<ipython-input-3-e36fbfe450f5> in <module>
----> 1 orb.oo_propagate(Time.now())

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/sbpy/data/orbit.py in oo_propagate(self, epoch, timescale, dynmodel, ephfile)
    544
    545         oo_orbits, err = pyoorb.pyoorb.oorb_propagation(
--> 546             in_orbits=self._to_oo(),
    547             in_epoch=ooepoch,
    548             in_dynmodel=dynmodel)

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/sbpy/data/orbit.py in _to_oo(self, timescale)
    234         if timescale is None:
    235             timescale_ = [conf.oorb_timeScales[t]
--> 236                           for t in self.table['timescale']]
    237         else:
    238             timescale_ = [conf.oorb_timeScales[timescale]] * len(self.table)

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/astropy/table/table.py in __getitem__(self, item)
   1223     def __getitem__(self, item):
   1224         if isinstance(item, str):
-> 1225             return self.columns[item]
   1226         elif isinstance(item, (int, np.integer)):
   1227             return self.Row(self, item)

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/astropy/table/table.py in __getitem__(self, item)
    105         """
    106         if isinstance(item, str):
--> 107             return OrderedDict.__getitem__(self, item)
    108         elif isinstance(item, (int, np.integer)):
    109             return self.values()[item]

KeyError: 'timescale'
```

I then tried to replace the plain `Quantity` in the dict with a proper `Time` object without success:

```
In [8]: elements["epoch"] = Time(2451200.5 * u.d, format="jd")

In [9]: elements
Out[9]:
{'a': <Quantity 1.234 AU>,
 'e': 0.1234,
 'i': <Quantity 12.34 deg>,
 'argper': <Quantity 123.4 deg>,
 'node': <Quantity 45.2 deg>,
 'epoch': <Time object: scale='utc' format='jd' value=2451200.5>,
 'true_anom': <Quantity 23.1 deg>}

In [10]: orb = Orbit.from_dict(elements)

In [11]: orb
Out[11]:
<QTable length=1>
   a       e       i     argper   node    epoch   true_anom
   AU             deg     deg     deg                deg
float64 float64 float64 float64 float64   object   float64
------- ------- ------- ------- ------- --------- ---------
  1.234  0.1234   12.34   123.4    45.2 2451200.5      23.1

In [12]: orb.oo_propagate(Time.now())
KeyError                                  Traceback (most recent call last)
<ipython-input-12-e36fbfe450f5> in <module>
----> 1 orb.oo_propagate(Time.now())

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/sbpy/data/orbit.py in oo_propagate(self, epoch, timescale, dynmodel, ephfile)
    544
    545         oo_orbits, err = pyoorb.pyoorb.oorb_propagation(
--> 546             in_orbits=self._to_oo(),
    547             in_epoch=ooepoch,
    548             in_dynmodel=dynmodel)

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/sbpy/data/orbit.py in _to_oo(self, timescale)
    234         if timescale is None:
    235             timescale_ = [conf.oorb_timeScales[t]
--> 236                           for t in self.table['timescale']]
    237         else:
    238             timescale_ = [conf.oorb_timeScales[timescale]] * len(self.table)

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/astropy/table/table.py in __getitem__(self, item)
   1223     def __getitem__(self, item):
   1224         if isinstance(item, str):
-> 1225             return self.columns[item]
   1226         elif isinstance(item, (int, np.integer)):
   1227             return self.Row(self, item)

~/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/astropy/table/table.py in __getitem__(self, item)
    105         """
    106         if isinstance(item, str):
--> 107             return OrderedDict.__getitem__(self, item)
    108         elif isinstance(item, (int, np.integer)):
    109             return self.values()[item]

KeyError: 'timescale'
```
