### 2019-06-09 NASA-Planetary-Science/sbpy#150 @Juanlu001
Testing requires extra steps
I tried testing the package after `pip install pytest`, [the only testing dependency listed in the docs](https://sbpy.readthedocs.io/en/latest/#requirements), and got an error:

```
(sbpy37) juanlu@centauri /tmp/sbpy $ pytest -v
============================================= test session starts ==============================================
platform linux -- Python 3.7.3, pytest-4.6.2, py-1.8.0, pluggy-0.12.0 -- /home/juanlu/.miniconda36/envs/sbpy37/bin/python
cachedir: .pytest_cache
rootdir: /tmp/sbpy, inifile: setup.cfg
collecting ... /home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/mark/structures.py:337: PytestUnknownMarkWarning: Unknown pytest.mark.remote_data - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/latest/mark.html
  PytestUnknownMarkWarning,
collected 0 items / 1 errors

==================================================== ERRORS ====================================================
________________________________________ ERROR collecting test session _________________________________________
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/config/__init__.py:440: in _importconftest
    return self._conftestpath2mod[conftestpath]
E   KeyError: local('/tmp/sbpy/sbpy/conftest.py')

/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/config/__init__.py:446: in _importconftest
    mod = conftestpath.pyimport()
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/py/_path/local.py:701: in pyimport
    __import__(modname)
<frozen importlib._bootstrap>:983: in _find_and_load
    ???
<frozen importlib._bootstrap>:967: in _find_and_load_unlocked
    ???
<frozen importlib._bootstrap>:668: in _load_unlocked
    ???
<frozen importlib._bootstrap>:638: in _load_backward_compatible
    ???
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/assertion/rewrite.py:304: in load_module
    exec(co, mod.__dict__)
sbpy/conftest.py:47: in <module>
    TESTED_VERSIONS['astropy_helpers'] = astropy_helpers_version
E   NameError: name 'astropy_helpers_version' is not defined

/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/py/_path/common.py:377: in visit
    for x in Visitor(fil, rec, ignore, bf, sort).gen(self):
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/py/_path/common.py:418: in gen
    dirs = self.optsort([p for p in entries
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/py/_path/common.py:419: in <listcomp>
    if p.check(dir=1) and (rec is None or rec(p))])
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/main.py:661: in _recurse
    ihook = self.gethookproxy(dirpath)
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/main.py:482: in gethookproxy
    my_conftestmodules = pm._getconftestmodules(fspath)
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/config/__init__.py:424: in _getconftestmodules
    mod = self._importconftest(conftestpath.realpath())
/home/juanlu/.miniconda36/envs/sbpy37/lib/python3.7/site-packages/_pytest/config/__init__.py:463: in _importconftest
    raise ConftestImportFailure(conftestpath, sys.exc_info())
E   _pytest.config.ConftestImportFailure: (local('/tmp/sbpy/sbpy/conftest.py'), (<class 'NameError'>, NameError("name 'astropy_helpers_version' is not defined"), <traceback object at 0x7f5286892948>))
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
=========================================== 1 error in 0.61 seconds ============================================
```

This code is likely the cause:

https://github.com/NASA-Planetary-Science/sbpy/blob/9867a8622b0034a0e042a42e303f80d566eed749/sbpy/conftest.py#L40-L47
