### 2019-06-09 NASA-Planetary-Science/sbpy#151 @Juanlu001
setup.py should not manually read metadata from setup.cfg
See https://setuptools.readthedocs.io/en/latest/setuptools.html?highlight=setup.cfg#configuring-setup-using-setup-cfg-files

For example: https://github.com/poliastro/poliastro/blob/8c3fdfcfbc02be18ce3c9b7a1b6703a818d1288a/setup.py#L8
