### 2019-06-09 NASA-Planetary-Science/sbpy#152 @Juanlu001
Tag and version mismatch
A `v0.1` tag exists, however

https://github.com/NASA-Planetary-Science/sbpy/blob/v0.1/setup.cfg#L35
