### 2019-06-09 NASA-Planetary-Science/sbpy#152 @Juanlu001
That's probably the reason why the docs still point to `v0.0.dev604`
