### 2019-06-09 #1426 @Juanlu001
Sorry all for the delay! I finished the review, having paid more attention to the orbital stuff and installation and testing procedures. Some comments:

* There seems to be [a v0.1 release](https://github.com/NASA-Planetary-Science/sbpy/releases/tag/v0.1) but [sbpy does not exist in PyPI](https://pypi.org/search/?q=sbpy). Is this intentional?
* https://github.com/NASA-Planetary-Science/sbpy/issues/151
* https://github.com/NASA-Planetary-Science/sbpy/issues/152
* https://github.com/NASA-Planetary-Science/sbpy/issues/149
* It's not entirely clear to me what's the difference between `from_horizons` and `from_mpc` in `Orbit` and `Ephem` objects. Or why there is a `Phys.from_sbdb`, but not a `Orbit.from_sbdb` (or `Ephem.from_sbdb`).
* `Orbit` methods return data in a coordinate system that is not made explicit, which can be a source of confusion. `Ephem` methods have a similar issue although I would need to look carefully at all the columns returned.
* The narrative documentation of `imageanalysis` and `thermal` are noticeably shorter than the other modules, and API docs of the latter are a bit lacking (except for `ThermalClass`)
* There _are_ automated tests, but https://github.com/NASA-Planetary-Science/sbpy/issues/150
* Also, it would be nice to describe these steps, even if they are obvious to experienced developers, in https://sbpy.readthedocs.io/en/latest/contributing.html

I left some boxes unchecked until the original authors clarify the status of the `v0.1` version, the testing issues, and the API docs of `sbpy.thermal`.
