### 2019-06-11 #1426 @mommermi
Thank, @Juanlu001, for your review! I will work my way through it and hopefully finish it this week.
