### 2019-06-11 NASA-Planetary-Science/sbpy#153 @mommermi
Improve documentation for data classes
* define differences between `Ephem`, `Orbit`, `Phys`
* explain available services
