### 2019-06-11 NASA-Planetary-Science/sbpy#154 @mommermi
identify coordinate systems in Orbit and Ephem objects
clarify the coordinate systems used in `Orbit` and `Ephem` objects
