### 2019-06-11 NASA-Planetary-Science/sbpy#155 @mommermi
improve documentation for contributions
it would be nice to describe these steps, even if they are obvious to experienced developers, in https://sbpy.readthedocs.io/en/latest/contributing.html
