### 2019-06-11 #1426 @mommermi
Here are some replies to you general comments:

> There seems to be a v0.1 release but sbpy does not exist in PyPI. Is this intentional?

No, we simply haven't gotten around to register it with PyPI. Most likely we will wait for this until v0.2 is released in a few weeks.

> It's not entirely clear to me what's the difference between from_horizons and from_mpc in Orbit and Ephem objects. Or why there is a Phys.from_sbdb, but not a Orbit.from_sbdb (or Ephem.from_sbdb).

`Ephem` objects are intended to hold ephemeris data, or, more generally, data that are time-dependent. `Orbit` objects on the other side hold orbital elements. [SBDB](https://ssd.jpl.nasa.gov/sbdb.cgi) is only used to extract physical properties - although it could be used to extract orbital elements, as well. We made this decision as `Orbit.from_horizons` taps the same data source including some additional information.

* [ ]  I already have an item on my to-do list to better explain the difference between `Ephem`, `Orbit`, and `Phys` in the docs, which should clarify the differences and why the different services have been implemented in these classes.  I added this issue as NASA-Planetary-Science/sbpy#153

> Orbit methods return data in a coordinate system that is not made explicit, which can be a source of confusion. Ephem methods have a similar issue although I would need to look carefully at all the columns returned.

* [ ] I will clarify this: NASA-Planetary-Science/sbpy#154

> The narrative documentation of imageanalysis and thermal are noticeably shorter than the other modules, and API docs of the latter are a bit lacking (except for ThermalClass)

The documentation for some of the modules (including `imageanalysis` and `thermal`) are simply skeletons as they have not yet been fully implemented. Documentation will be improved and added once the modules have been further implemented.

> Also, it would be nice to describe these steps, even if they are obvious to experienced developers, in https://sbpy.readthedocs.io/en/latest/contributing.html

* [ ] Will do: NASA-Planetary-Science/sbpy#155
