### 2019-06-12 NASA-Planetary-Science/sbpy#155 @bsipocz
You're welcome to reuse the astropy contribution guide (in case you follow that, or just gut out the things you don't need). In e.g. astroquery we just list what we do on top of that guide (see e.g https://github.com/astropy/astroquery/blob/master/CONTRIBUTING.rst)
