### 2019-06-12 NASA-Planetary-Science/sbpy#151 @bsipocz
@Juanlu001 - sbpy uses the astropy package template, including the setup files and as such comes with a lot of historical etc baggage (e.g. supporting older setuptools versions, providing a template to support both python2 and python3 versions, etc), yet for overall maintenance it would be easier for everyone involved to minimize divergence from the template unless there are good reasons for it. Thus I would suggest removing this item from the list of issues for the joss review.

As of the package template, we're about to tag a new version that has reworked the setup infrastructure and ultimately makes setuptools 30.3 the minimum required version.
