### 2019-06-12 NASA-Planetary-Science/sbpy#150 @bsipocz
@Juanlu001 - While it's indeed not mentioned in the documentation, the package uses the astropy testing infrastructure thus it should be run via setup.py  to pick up all the configurations correctly:
`python setup.py test`
