### 2019-06-12 NASA-Planetary-Science/sbpy#150 @bsipocz
e.g. when you tried I suppose the submodule hasn't been initialized (the astropy testrunner does initialize it, but pytest doesn't) and thus the `NameError` you got. Once you properly have the submodule `pytest sbpy` should work as expected.
