### 2019-06-12 NASA-Planetary-Science/sbpy#151 @Juanlu001
Fair enough! Let's not make this a blocker. Looking forward to the new template :)
