### 2019-06-12 6324024
added additional alternative field names
