+++ sbpy/data/ephem.py
@@ -62,6 +62,16 @@ class Ephem(DataClass):
             Arguments that will be provided to
             `astroquery.jplhorizons.HorizonsClass.ephemerides`.
 
+        Notes
+        -----
+        * For detailed explanations of the queried fields, refer to
+          `astroquery.jplhorizons.HorizonsClass.ephemerides <https://astroquery.readthedocs.io/en/latest/api/astroquery.jplhorizons.HorizonsClass.html#astroquery.jplhorizons.HorizonsClass.ephemerides>`_ and the
+          `JPL Horizons documentation <https://ssd.jpl.nasa.gov/?horizons_doc>`_.
+        * By default, all properties are provided in the J2000.0 reference
+          system. Different settings can be chosen using
+          additional keyword arguments as used by
+          `astroquery.jplhorizons.HorizonsClass.ephemerides <https://astroquery.readthedocs.io/en/latest/api/astroquery.jplhorizons.HorizonsClass.html#astroquery.jplhorizons.HorizonsClass.ephemerides>`_.
+
         Returns
         -------
         `~Ephem` object
