### 2019-06-12 NASA-Planetary-Science/sbpy#154 @mommermi
I added some links to the `from_horizons` docstring to clarify the default reference system for both orbital elements and ephemerides, as well as a short explanation on how to change them. The same will be done for `from_mpc` and other query wrappers as soon as they are implemented. I also added links that point to field name definitions for properties that are queried.

I chose this approach over explicitly adding this information to the query results as that approach would lead to a large overhead (e.g., adding a `refsys`/`refplane` column to the corresponding object would only bloat the underlying table object).

changes are commited here: 4d616d1
