### 2019-06-12 980adbb
added more details on data containers
