### 2019-06-12 NASA-Planetary-Science/sbpy#153 @mommermi
more details have been added here: 980adbb
