### 2019-06-12 NASA-Planetary-Science/sbpy#149 @mommermi
This issue should have been taken care of by 4d616d1. The problem was that the `orb` object as defined above was not complete for use in `orb.oo_propagate`. Two fields were missing in `orb`: `timescale` and `targetname`. Both properties are now filled in automatically (`timescale` is set to UTC) if not already available in the `Orbit` object. This should make the openorb functionality a lot more accessible to users. Thanks for pointing this out!
