### 2019-06-12 NASA-Planetary-Science/sbpy#151 @mommermi
Thanks, @bsipocz  and @Juanlu001!
