### 2019-06-12 NASA-Planetary-Science/sbpy#152 @mommermi
this has been fixed here: f151802
