### 2019-06-12 NASA-Planetary-Science/sbpy#155 @mommermi
Perfect, thanks, @bsipocz!
