### 2019-06-12 b8c9a84
code contribution docs updated
