### 2019-06-12 NASA-Planetary-Science/sbpy#150 @mommermi
I added a note in b8c9a84. referencing the astropy testing guidelines.
