### 2019-06-12 NASA-Planetary-Science/sbpy#155 @mommermi
I updated the guidelines here: b8c9a84

Please let me know if you suggest anything else to be mentioned here.
