### 2019-06-12 #1426 @mommermi
@Juanlu001 I think I implemented all the changes. My replies and corresponding changes in the code are linked in the corresponding issues:

NASA-Planetary-Science/sbpy#149
NASA-Planetary-Science/sbpy#150
NASA-Planetary-Science/sbpy#152
NASA-Planetary-Science/sbpy#153
NASA-Planetary-Science/sbpy#154
NASA-Planetary-Science/sbpy#155

NASA-Planetary-Science/sbpy#151 has already been closed.

If there is anything else I can do, please let me know!
