### 2019-06-13 7ef12bc
root-level information for contributions and code0of-conduct
