### 2019-06-13 a9e1ced
added root-level docs for contributions and code-of-conduct
