+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,3 @@
+sbpy follows the Astropy code of conduct. 
+
+All Astropy community members are expected to abide by the [Astropy Project Code of Conduct](http://www.astropy.org/code_of_conduct.html).
