### 2019-06-14 #1426 @Juanlu001
Excellent @mommermi! I checked all the boxes in the review and think this is good to go with https://github.com/NASA-Planetary-Science/sbpy/pull/156.
