### 2019-06-14 #1426 @mommermi
Thanks, @Juanlu001! I will wait with merging NASA-Planetary-Science/sbpy#156 until the astropy core issue is fixed and the CI builds succeed again. Thanks for your detailed review!
