### 2019-06-17 #1426 @xuanxu
Everything looks good. @mommermi After the revision changes version 0.1 is 27 commits behind master. Could you please release a new version from the current master?

Once you do that please create a Zenodo archive of that new version and report its DOI here (edit the Zenodo deposit metadata (title and author list) to match [the paper](https://github.com/openjournals/joss-papers/blob/joss.01426/joss.01426/10.21105.joss.01426.pdf)).
