### 2019-06-17 3bb4c92
update version number for JOSS publication
