+++ setup.cfg
@@ -32,7 +32,7 @@ url = http://sbpy.org
 edit_on_github = False
 github_project = NASA-Planetary-Science/sbpy
 # version should be PEP386 compatible (http://www.python.org/dev/peps/pep-0386)
-version = 0.1.dev
+version = 0.1.1.dev
 install_requires = numpy,astropy-helpers,matplotlib,ads,synphot,astroquery
 
 [pycodestyle]
