### 2019-06-17 #1426 @mommermi
Thanks, @xuanxu, it's in the works. I have to wait for our NASA program officer (who is in charge of the github organization under which sbpy resides) to enable third-party access to the repo for zenodo. I'll ping you once we have the doi.
