### 2019-06-18 a9c891f
DataClass returns QTable instead of Row
