### 2019-06-21 #1426 @mommermi
Ok, the latest version is registered with zenodo. The DOI is 10.5281/zenodo.3252172
