### 2019-06-22 #1426 @xuanxu
Yay! :tada:
Everything is ready por publication. Pinging @openjournals/joss-eics for final acceptance.
