### 2019-06-22 #1426 @danielskatz
@xuanxu - there are 2 typos in the papers, as indicated in https://github.com/NASA-Planetary-Science/sbpy/pull/158

Once these are fixed (by merging this PR), we can finish the acceptance
