### 2019-06-22 #1426 @danielskatz
@mommermi - Can you merge https://github.com/NASA-Planetary-Science/sbpy/pull/158 ?
