### 2019-06-22 #1426 @danielskatz
Thanks to @Juanlu001 for reviewing and @xuanxu for editing
